fix hooks used outside of hook or component
