+++ next/components/molecules/SearchPageNew/SearchCardComposed.tsx
@@ -4,7 +4,7 @@ import { Typography } from '@bratislava/component-library'
 import ImagePlaceholder from '@components/atoms/ImagePlaceholder'
 import MLink from '@components/forms/simple-components/MLink'
 import Tag from '@components/forms/simple-components/Tag'
-import { SearchResult } from '@components/molecules/SearchPageNew/searchDataFetchers'
+import { SearchResult } from '@components/molecules/SearchPageNew/useQueryBySearchOption'
 import { getCategoryColorLocalStyle } from '@utils/colors'
 import { generateImageSizes } from '@utils/generateImageSizes'
 import { isDefined } from '@utils/isDefined'
