+++ next/messages/en.json
@@ -259,6 +259,7 @@
     "allResults": "All results",
     "moreResults": "More results",
     "noResults": "No results found",
+    "enterSearchQuery": "Enter a search query",
     "article": "Article",
     "articles": "Articles",
     "inbaArticle": "in.ba article",
