+++ next/messages/sk.json
@@ -259,6 +259,7 @@
     "allResults": "Všetky výsledky",
     "moreResults": "Viac výsledkov",
     "noResults": "Žiadne výsledky",
+    "enterSearchQuery": "Zadajte hľadaný výraz",
     "article": "Článok",
     "articles": "Články",
     "inbaArticle": "in.ba článok",
